test: consolidate api tests into tests/test_ai_api.py, remove orphan ai/test_api.py
